requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ase==3.21.1
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blis==0.7.9
catalogue==2.0.8
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.6
cycler==0.10.0
cymem==2.0.7
datasets==2.7.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.3
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.31
googledrivedownloader==0.4
huggingface-hub==0.11.1
hydra-core==1.3.1
idna==2.10
imageio==2.25.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
langcodes==3.3.0
lark==1.1.5
llvmlite==0.39.1
MarkupSafe==1.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.3
multiprocess==0.70.14
murmurhash==1.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.2.0
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
packaging==20.9
pandas==1.5.2
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.0
ply==3.11
preshed==3.0.8
progressbar==2.5
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==10.0.1
pycparser==2.20
pydantic==1.10.2
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-louvain==0.16
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rdflib==5.0.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
sip==6.7.5
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
spacy-pkuseg==0.0.32
srsly==2.4.5
sympy==1.11.1
terminado==0.17.1
testpath==0.4.4
thinc==8.1.6
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tokenizers==0.13.2
toml==0.10.2
torch==1.13.0
torch-explain==1.3.0
torch-geometric==1.7.2
torch-scatter==2.1.0
torch-sparse @ git+https://github.com/rusty1s/pytorch_sparse.git@955b1cf31ac0e12027dbf3008770e494ac5efd9c
torchaudio==0.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.13
wandb==0.13.10
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.4
xxhash==3.1.0
yarl==1.8.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.4.0/zh_core_web_sm-3.4.0-py3-none-any.whl
